mod config instead of enabler
