+++ src/services/index.ts
@@ -1,6 +1,6 @@
 export * from './get-local-mods';
 export * from './get-mod-database';
-export * from './mod-enabler';
+export * from './mod-config';
 export * from './mod-manager';
 export * from './run-owml';
 export * from './self-update';
