+++ src/services/mod-config.ts
@@ -16,10 +16,6 @@ function isConfigExisting(mod: Mod) {
   return fs.existsSync(getConfigPath(mod));
 }
 
-function saveConfig(mod: Mod, config: ModConfig) {
-  fs.writeJSONSync(getConfigPath(mod), config);
-}
-
 function getDefaultConfig(mod: Mod): ModConfig {
   if (isDefaultConfigExisting(mod)) {
     const defaultConfig: ModConfig = fs.readJSONSync(getDefaultConfigPath(mod));
@@ -32,21 +28,14 @@ function getDefaultConfig(mod: Mod): ModConfig {
   }
 }
 
-function getConfig(mod: Mod): ModConfig {
+export function saveConfig(mod: Mod, config: ModConfig) {
+  fs.writeJSONSync(getConfigPath(mod), config);
+}
+
+export function getConfig(mod: Mod): ModConfig {
   if (!isConfigExisting(mod)) {
     const defaultConfig = getDefaultConfig(mod);
     saveConfig(mod, defaultConfig);
   }
   return fs.readJSONSync(getConfigPath(mod));
-}
-
-export function isEnabled(mod: Mod): boolean {
-  const config = getConfig(mod);
-  return config.enabled;
-}
-
-export function toggleEnabled(mod: Mod) {
-  const config = getConfig(mod);
-  config.enabled = !config.enabled;
-  saveConfig(mod, config);
 }
